refactor: extract `HTTPView`.
